+++ jest.config.js
@@ -1,3 +1,7 @@
 import { jest } from "@snowcoders/renovate-config";
 
-export default jest;
+export default {
+  ...jest,
+  // We're running actual applications in this situation, don't want to run multiple at the same time
+  maxWorkers: 1,
+};