+++ src/cli.e2e.test.ts
@@ -1,52 +1,33 @@
-import { exec, execSync } from "node:child_process";
-import { describe, it, expect } from "@jest/globals";
+import { expect, it } from "@jest/globals";
+import { execSync } from "node:child_process";
 
-type CommandResult = {
-  exitCode: number;
-  stdout: string;
-  stderr: string;
-};
-
-async function runCommand(command: string): Promise<CommandResult> {
-  return new Promise<CommandResult>((resolve, reject) => {
-    try {
-      exec(
-        command,
-        {
-          cwd: process.cwd(),
-          env: process.env,
-          timeout: 5000,
-        },
-        (error, stdout, stderr) => {
-          return resolve({
-            exitCode: error?.code ?? 0,
-            stderr: stderr.trim(),
-            stdout: stdout.trim(),
-          });
-        }
-      );
-    } catch (e) {
-      reject(e);
-    }
-  });
+function runCommand(commandString: string): number {
+  const command = commandString.split(" ");
+  try {
+    execSync(commandString, {
+      cwd: process.cwd(),
+      env: process.env,
+      stdio: "ignore",
+    });
+    return 0;
+  } catch (e) {
+    // @ts-ignore
+    return e.status;
+  }
 }
 
-it.each`
-  command
-  ${"eslint --kjhygjkhg"}
-  ${"prettier -h"}
-`("Returns the same exit code for command $command", async ({ command }) => {
-  const expected = await runCommand(`npm exec --no -- ${command}`);
-  const actualRun = await runCommand(`node ./dist/cli.js ${command}`);
-  expect(expected).toEqual(actualRun);
-});
+it.each<{ command: string; expectedExitCode: number; notes: string }>`
+  command                                      | expectedExitCode | notes
+  ${"npm"}                                     | ${0}             | ${"installed globally but not locally"}
+  ${"does-not-exist"}                          | ${0}             | ${"not installed and is not a real package"}
+  ${"esbuild"}                                 | ${0}             | ${"not installed however is an actual package"}
+  ${"eslint --kjhygjkhg"}                      | ${2}             | ${"installed locally with invalid arguments"}
+  ${"prettier"}                                | ${2}             | ${"installed locally with no arguments (as of 2.8.3)"}
+  ${"prettier -- this-file-does-not-exist.js"} | ${2}             | ${"installed locally with invalid file to run (as of 2.8.3)"}
+  ${"prettier --flag-does-not-exist"}          | ${2}             | ${"installed locally with invalid arguments (as of 2.8.3)"}
+  ${"prettier -v"}                             | ${0}             | ${"installed locally with valid arguments (as of 2.8.3)"}
+`("Returns $expectedExitCode if program is $notes: '$command'", ({ command, expectedExitCode }) => {
+  const execIfExistsExitCode = runCommand(`node ./dist/cli.js ${command}`);
 
-it.each`
-  command
-  ${"does-not-exist"}
-`("Returns exit code 0 for command 'does-not-exist'", async ({ command }) => {
-  const expected = await runCommand(`npm exec --no -- ${command}`);
-  const actualRun = await runCommand(`node ./dist/cli.js ${command}`);
-  expect(expected.exitCode).toBe(1);
-  expect(actualRun.exitCode).toBe(0);
+  expect(execIfExistsExitCode).toBe(expectedExitCode);
 });