+++ src/index.test.ts
@@ -22,16 +22,16 @@ describe("Error scenarios", () => {
   });
 });
 
-describe("Success scenarios", () => {
-  it("Returns 0 if program is not found", async () => {
-    await expect(execIfExists(["does-not-exist"])).resolves.toBe(0);
-  });
-
-  it("Returns 0 on successful execution", async () => {
-    await expect(execIfExists(["npm", "-v"])).resolves.toBe(0);
-  });
-
-  it("Returns not 0 on non-successful execution", async () => {
-    await expect(execIfExists(["npm", "--flag-does-not-exist"])).resolves.not.toBe(0);
-  });
+it.each<{ command: string; expectedExitCode: number; notes: string }>`
+  command                                      | expectedExitCode | notes
+  ${"npm"}                                     | ${0}             | ${"installed globally but not locally"}
+  ${"does-not-exist"}                          | ${0}             | ${"not installed and is not a real package"}
+  ${"esbuild"}                                 | ${0}             | ${"not installed however is an actual package"}
+  ${"eslint --kjhygjkhg"}                      | ${2}             | ${"installed locally with invalid arguments"}
+  ${"prettier"}                                | ${2}             | ${"installed locally with no arguments (as of 2.8.3)"}
+  ${"prettier -- this-file-does-not-exist.js"} | ${2}             | ${"installed locally with invalid file to run (as of 2.8.3)"}
+  ${"prettier --flag-does-not-exist"}          | ${2}             | ${"installed locally with invalid arguments (as of 2.8.3)"}
+  ${"prettier -v"}                             | ${0}             | ${"installed locally with valid arguments (as of 2.8.3)"}
+`("Returns $expectedExitCode if program is $notes: '$command'", async ({ command, expectedExitCode }) => {
+  await expect(execIfExists(command.split(" "))).resolves.toBe(expectedExitCode);
 });