+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed package exists in registry but not on local install
+
 ## [4.0.0-beta.1] - 2023-01-16
 
 - No longer hides underlying commands output unless it results in a E404